# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CreviaTube API
  description: |
    Advanced content creator platform API for monetizing content and managing campaigns.
    
    ## Authentication
    Most endpoints require authentication. Include your session cookie or use the login endpoint to authenticate.
    
    ## Rate Limiting
    - Authentication endpoints: 5 attempts per 15 minutes
    - API endpoints: 100 requests per 15 minutes
    - Sensitive operations: 10 requests per hour
    - Tracking events: 30 events per minute
    - Admin operations: 50 requests per 15 minutes
    
    ## Error Codes
    - `400` - Bad Request
    - `401` - Unauthorized
    - `403` - Forbidden
    - `404` - Not Found
    - `429` - Too Many Requests
    - `500` - Internal Server Error
  version: 1.0.0
  contact:
    name: CreviaTube Support
    email: support@creviatube.com
    url: https://creviatube.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.creviatube.com
    description: Production server
security:
  - sessionAuth: []
paths:
...
  /api/analytics/stats:
    get:
      tags:
        - Analytics
      summary: Get analytics statistics
      description: Retrieve overall analytics statistics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Analytics statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsStats'
  /api/analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get analytics dashboard
      description: Retrieve analytics dashboard data
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [overview, funnels, cohorts, geographic, predictions]
            default: overview
          description: Dashboard type
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month, year]
            default: month
          description: Time period
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Dashboard data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardData'

  # ===== VISUALIZATION ENDPOINTS =====
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
...